# examples/reference-app/internal/cases/api/cases-server.gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

// CasesCase A case as the create answer, the list entries and the detail answer all render it -- one shape a P3 view renders with one component.
type CasesCase struct {
	// CreatedAt When the case was created, whole seconds, UTC.
	CreatedAt time.Time `json:"created_at"`

	// CreatorUserID The user id of the staff member who created the case, as the host's SubjectResolver seam attributed the create request. Recorded on the row for the case's history; the clinic's case list is NOT filtered by it (every member of the tenant sees every case of the tenant).
	CreatorUserID string `json:"creator_user_id"`

	// ID The case's application-generated id.
	ID string `json:"id"`

	// PatientName The patient's display name as the clinic entered it.
	PatientName string `json:"patient_name"`

	// PatientRef The clinic-given patient reference; empty means the clinic gave none.
	PatientRef string `json:"patient_ref"`

	// Photos The case's photos in attachment order. Always present -- an empty list for a case whose photos arrive later.
	Photos []CasesPhoto `json:"photos"`
}

// CasesCreateCaseRequest The create route's request body.
type CasesCreateCaseRequest struct {
	// PatientName The patient's display name as the clinic entered it at intake. Required, length-bounded in the application (SQLite does not enforce a VARCHAR length limit, so the check lives in the service).
	PatientName string `json:"patient_name"`

	// PatientRef The optional clinic-given identifier for the patient (a chart number, an internal reference). An absent or empty value means the clinic gave none.
	PatientRef *string `json:"patient_ref,omitempty"`

	// PhotoObjectIds The case's initial photos, in attachment order: references to already-uploaded go/storage objects of the caller's tenant. Optional -- an absent or empty list creates a case with no photos yet. Each entry must be a non-empty object id, unique within the list and not already attached to another case of the same tenant.
	PhotoObjectIds *[]string `json:"photo_object_ids,omitempty"`
}

// CasesError The structured {code, params} error envelope every speed API returns instead of localized text: code is a stable machine-readable identifier naming exactly one failure, and params carries the structured per-code details when the failure has any. The server never renders a message -- a client maps each code to its own locale's text.
type CasesError struct {
	// Code Example: cases.patient_name_required
	Code   *string                 `json:"code,omitempty"`
	Params *map[string]interface{} `json:"params,omitempty"`
}

// CasesListResponse The list route's 200 answer.
type CasesListResponse struct {
	// Cases The caller's tenant's cases, newest first.
	Cases []CasesCase `json:"cases"`
}

// CasesPhoto One photo of a case, as a case read answers it.
type CasesPhoto struct {
	// ObjectID The referenced go/storage photo object's id -- the value that feeds the smile-simulation surface's per-photo enumeration.
	ObjectID string `json:"object_id"`
}

// CasesPhotoContent The photo-content route's 200 answer.
type CasesPhotoContent struct {
	// ContentBase64 The photo's stored bytes, base64-encoded -- the value a case view decodes into a blob URL it renders.
	ContentBase64 string `json:"content_base64"`

	// MediaType The media type storage's probe assigned to the bytes at complete time (the sanitized form), e.g. image/jpeg.
	MediaType string `json:"media_type"`
}

// CasesUploadPhotoRequest The photo-upload route's request body.
type CasesUploadPhotoRequest struct {
	// ContentBase64 The photo's bytes, base64-encoded (standard alphabet with padding, as the browser's FileReader data URL carries them). The decoded bytes must be non-empty and at most the surface's upload bound; storage's own probe of the stored bytes decides whether they are an acceptable image.
	ContentBase64 string `json:"content_base64"`
}

// CasesUploadPhotoResponse The photo-upload route's 201 answer.
type CasesUploadPhotoResponse struct {
	// ObjectID The completed go/storage photo object's id -- the value a case create's photo_object_ids list references to attach the photo to a case.
	ObjectID string `json:"object_id"`
}

// CasesCreateCaseJSONRequestBody defines body for CasesCreateCase for application/json ContentType.
type CasesCreateCaseJSONRequestBody = CasesCreateCaseRequest

// CasesUploadPhotoJSONRequestBody defines body for CasesUploadPhoto for application/json ContentType.
type CasesUploadPhotoJSONRequestBody = CasesUploadPhotoRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// CasesListCases List every case of the caller's tenant.
	// (GET /api/v1/cases)
	CasesListCases(w http.ResponseWriter, r *http.Request)
	// CasesCreateCase Create a case under the caller's tenant.
	// (POST /api/v1/cases)
	CasesCreateCase(w http.ResponseWriter, r *http.Request)
	// CasesUploadPhoto Upload one patient photo to the caller's tenant.
	// (POST /api/v1/cases/photos/upload)
	CasesUploadPhoto(w http.ResponseWriter, r *http.Request)
	// CasesGetCase Read one case of the caller's tenant.
	// (GET /api/v1/cases/{caseId})
	CasesGetCase(w http.ResponseWriter, r *http.Request, caseID string)
	// CasesGetPhotoContent Read one photo of one case of the caller's tenant.
	// (GET /api/v1/cases/{caseId}/photos/{photoObjectID}/content)
	CasesGetPhotoContent(w http.ResponseWriter, r *http.Request, caseID string, photoObjectID string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// CasesListCases operation middleware
func (siw *ServerInterfaceWrapper) CasesListCases(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CasesListCases(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CasesCreateCase operation middleware
func (siw *ServerInterfaceWrapper) CasesCreateCase(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CasesCreateCase(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CasesUploadPhoto operation middleware
func (siw *ServerInterfaceWrapper) CasesUploadPhoto(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CasesUploadPhoto(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CasesGetCase operation middleware
func (siw *ServerInterfaceWrapper) CasesGetCase(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "caseId" -------------
	var caseID string

	err = runtime.BindStyledParameterWithOptions("simple", "caseId", r.PathValue("caseId"), &caseID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "caseId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CasesGetCase(w, r, caseID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CasesGetPhotoContent operation middleware
func (siw *ServerInterfaceWrapper) CasesGetPhotoContent(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "caseId" -------------
	var caseID string

	err = runtime.BindStyledParameterWithOptions("simple", "caseId", r.PathValue("caseId"), &caseID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "caseId", Err: err})
		return
	}

	// ------------- Path parameter "photoObjectID" -------------
	var photoObjectID string

	err = runtime.BindStyledParameterWithOptions("simple", "photoObjectID", r.PathValue("photoObjectID"), &photoObjectID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "photoObjectID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CasesGetPhotoContent(w, r, caseID, photoObjectID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/v1/cases", wrapper.CasesListCases)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/v1/cases", wrapper.CasesCreateCase)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/v1/cases/photos/upload", wrapper.CasesUploadPhoto)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/v1/cases/{caseId}/photos/{photoObjectID}/content", wrapper.CasesGetPhotoContent)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/v1/cases/{caseId}", wrapper.CasesGetCase)

	return m
}