# examples/reference-app/web/app-openapi.yaml
openapi: 3.0.3
info:
  title: reference-app
  description: "The reference-app application OpenAPI document: the redocly join of the application own three fragments (notes, cases and smilesim), regenerated by `task api:gen:app` and committed. Each fragment stays authoritative for its own operations; this merged document is the app-owned SDK generation input and the linted whole-surface view."
  version: 0.0.0
...
paths:
...
  /api/v1/cases:
    servers: []
    post:
      operationId: cases_createCase
      tags:
        - cases
      summary: Create a case under the caller's tenant.
      description: >-
        The case groups one patient (the clinic-given name required, the
        optional clinic-given reference) with the photos of the case as
        already-uploaded go/storage object ids, in attachment order. Both the
        photo list and the reference may be absent or empty -- a case can be
        created for an intake whose photos arrive later. There is deliberately
        no tenant_id field and no creator field anywhere on the request: the
        tenant is the one tenancy.Middleware already resolved into the request
        context, and the creator comes from the host's SubjectResolver seam,
        never from the request (a request no resolver can attribute is refused
        with cases.subject_unresolved before the body is even read).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CasesCreateCaseRequest'
      responses:
        '201':
          description: The case was created, with its photos in attachment order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesCase'
        '400':
          description: >-
            The request body was missing or malformed, or a documented bound was
            violated (cases.invalid_request_body, cases.patient_name_required,
            cases.patient_name_too_long, cases.patient_ref_too_long,
            cases.photo_object_id_required, cases.photo_object_id_too_long,
            cases.duplicate_photo_object, cases.too_many_photos).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesError'
        '401':
          description: >-
            The request could not be attributed to a user
            (cases.subject_unresolved).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesError'
        '403':
          description: >-
            The request carried no tenant -- an anonymous request to a
            tenant-scoped route (tenancy.tenant_unresolved).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesError'
        '409':
          description: >-
            A named photo object is already attached to another case of the same
            tenant (cases.photo_already_attached).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesError'
        '500':
          description: An unexpected server failure (cases.internal_error).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesError'
    get:
      operationId: cases_listCases
      tags:
        - cases
      summary: List every case of the caller's tenant.
      description: >-
        Lists every case of the caller's tenant, newest first -- the
        clinic-scoped list: everyone who treats a patient together in one
        practice sees the same cases, and only the tenant boundary hides a case
        (proven by the cross-tenant refusal tests, never by the creator column).
        Each entry carries its photos in attachment order (an empty list for a
        case whose photos arrive later). The list needs no creator attribution
        at all: any authenticated member of the tenant may read it, and the
        creator a create request is attributed with is recorded on the case row,
        never used to hide or filter the clinic's list.
      responses:
        '200':
          description: The caller's tenant's cases, newest first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesListResponse'
        '403':
          description: >-
            The request carried no tenant -- an anonymous request to a
            tenant-scoped route (tenancy.tenant_unresolved).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesError'
        '500':
          description: An unexpected server failure (cases.internal_error).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesError'
  /api/v1/cases/photos/upload:
    servers: []
    post:
      operationId: cases_uploadPhoto
      tags:
        - cases
      summary: Upload one patient photo to the caller's tenant.
      description: >-
        Uploads one patient photo as a go/storage object of the caller's tenant
        and returns its object id -- the value a case create's photo_object_ids
        list then references. The operation is this surface's one-shot form of
        go/storage's three-step upload protocol: the handler runs
        create/upload/complete itself, computing the declared size and checksum
        from the decoded bytes (so no declared-vs-arrived mismatch is possible)
        and letting storage's own probe of the stored bytes decide the media
        type and refuse what is not an acceptable image. The bytes travel
        base64-encoded in JSON because this surface's frontend transport (the
        generated operations over @speed/api-client) is JSON-only; the upload
        page decodes nothing -- the browser file's bytes are base64-encoded
        once, client-side. An uploaded photo is NOT yet attached to any case:
        attachment happens when a case create names the object id, and the
        one-tenant-one-case-per-photo rule (cases.photo_already_attached) then
        applies like it does to any other object id. A photo never attached to a
        case is reclaimed by go/storage's own expiry sweep like any other
        object.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CasesUploadPhotoRequest'
      responses:
        '201':
          description: >-
            The photo was uploaded and completed; the answer names its object
            id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesUploadPhotoResponse'
        '400':
          description: >-
            The request body was missing or malformed, or a documented bound was
            violated (cases.invalid_request_body, cases.photo_content_required,
            cases.photo_content_invalid, cases.photo_content_too_large,
            cases.photo_rejected).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesError'
        '403':
          description: >-
            The request carried no tenant -- an anonymous request to a
            tenant-scoped route (tenancy.tenant_unresolved).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesError'
        '500':
          description: An unexpected server failure (cases.internal_error).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesError'
  /api/v1/cases/{caseId}/photos/{photoObjectID}/content:
    servers: []
    get:
      operationId: cases_getPhotoContent
      tags:
        - cases
      summary: Read one photo of one case of the caller's tenant.
      description: >-
        Reads one photo of one case of the caller's tenant -- the bytes of the
        go/storage object a case photo references, base64-encoded in JSON with
        the media type storage's probe assigned at complete time, so the case
        view can render it (a blob URL built from the decoded bytes). The photo
        is addressed by its object id -- the value the case detail's photos
        entries carry -- never by a position the client would have to guess. An
        unknown case id, and a case of another tenant, both answer
        cases.not_found; a photo object id the case does not carry -- or whose
        object no longer exists in storage -- answers cases.photo_not_found;
        both refusals are deliberately indistinguishable from the same answer an
        unrelated request would get, so a caller can never probe another
        tenant's attachments.
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
          description: The case's application-generated id.
        - name: photoObjectID
          in: path
          required: true
          schema:
            type: string
          description: >-
            The photo's go/storage object id, as the case detail's photos
            entries carry it.
      responses:
        '200':
          description: The photo's bytes, with the media type storage's probe assigned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesPhotoContent'
        '400':
          description: >-
            The photo's bytes exceed the bound the surface can serve in one
            answer (cases.photo_content_too_large).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesError'
        '403':
          description: >-
            The request carried no tenant -- an anonymous request to a
            tenant-scoped route (tenancy.tenant_unresolved).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesError'
        '404':
          description: >-
            No such case under the caller's tenant (cases.not_found), or no such
            photo on that case (cases.photo_not_found) -- the same answer for
            another tenant's case or photo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesError'
        '500':
          description: An unexpected server failure (cases.internal_error).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesError'
  /api/v1/cases/{caseId}:
    servers: []
    get:
      operationId: cases_getCase
      tags:
        - cases
      summary: Read one case of the caller's tenant.
      description: >-
        Reads one case of the caller's tenant with its photos in attachment
        order. An unknown case id, and a case of another tenant, both answer
        cases.not_found -- never a cross-tenant confirmation of a case's
        existence.
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
          description: The case's application-generated id.
      responses:
        '200':
          description: The case, with its photos in attachment order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesCase'
        '403':
          description: >-
            The request carried no tenant -- an anonymous request to a
            tenant-scoped route (tenancy.tenant_unresolved).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesError'
        '404':
          description: >-
            No such case under the caller's tenant (cases.not_found -- the same
            answer for another tenant's case id).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesError'
        '500':
          description: An unexpected server failure (cases.internal_error).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesError'
...
components:
  schemas:
...
    CasesCreateCaseRequest:
      type: object
      required:
        - patient_name
      properties:
        patient_name:
          type: string
          minLength: 1
          maxLength: 200
          description: >-
            The patient's display name as the clinic entered it at intake.
            Required, length-bounded in the application (SQLite does not enforce
            a VARCHAR length limit, so the check lives in the service).
        patient_ref:
          type: string
          maxLength: 64
          description: >-
            The optional clinic-given identifier for the patient (a chart
            number, an internal reference). An absent or empty value means the
            clinic gave none.
        photo_object_ids:
          type: array
          maxItems: 50
          items:
            type: string
            minLength: 1
          description: >-
            The case's initial photos, in attachment order: references to
            already-uploaded go/storage objects of the caller's tenant. Optional
            -- an absent or empty list creates a case with no photos yet. Each
            entry must be a non-empty object id, unique within the list and not
            already attached to another case of the same tenant.
      description: The create route's request body.
    CasesPhoto:
      type: object
      required:
        - object_id
      properties:
        object_id:
          type: string
          description: >-
            The referenced go/storage photo object's id -- the value that feeds
            the smile-simulation surface's per-photo enumeration.
      description: One photo of a case, as a case read answers it.
    CasesCase:
      type: object
      required:
        - id
        - patient_name
        - patient_ref
        - creator_user_id
        - created_at
        - photos
      properties:
        id:
          type: string
          description: The case's application-generated id.
        patient_name:
          type: string
          description: The patient's display name as the clinic entered it.
        patient_ref:
          type: string
          description: >-
            The clinic-given patient reference; empty means the clinic gave
            none.
        creator_user_id:
          type: string
          description: >-
            The user id of the staff member who created the case, as the host's
            SubjectResolver seam attributed the create request. Recorded on the
            row for the case's history; the clinic's case list is NOT filtered
            by it (every member of the tenant sees every case of the tenant).
        created_at:
          type: string
          format: date-time
          description: When the case was created, whole seconds, UTC.
        photos:
          type: array
          items:
            $ref: '#/components/schemas/CasesPhoto'
          description: >-
            The case's photos in attachment order. Always present -- an empty
            list for a case whose photos arrive later.
      description: >-
        A case as the create answer, the list entries and the detail answer all
        render it -- one shape a P3 view renders with one component.
    CasesListResponse:
      type: object
      required:
        - cases
      properties:
        cases:
          type: array
          items:
            $ref: '#/components/schemas/CasesCase'
          description: The caller's tenant's cases, newest first.
      description: The list route's 200 answer.
    CasesUploadPhotoRequest:
      type: object
      required:
        - content_base64
      properties:
        content_base64:
          type: string
          minLength: 1
          description: >-
            The photo's bytes, base64-encoded (standard alphabet with padding,
            as the browser's FileReader data URL carries them). The decoded
            bytes must be non-empty and at most the surface's upload bound;
            storage's own probe of the stored bytes decides whether they are an
            acceptable image.
      description: The photo-upload route's request body.
    CasesUploadPhotoResponse:
      type: object
      required:
        - object_id
      properties:
        object_id:
          type: string
          description: >-
            The completed go/storage photo object's id -- the value a case
            create's photo_object_ids list references to attach the photo to a
            case.
      description: The photo-upload route's 201 answer.
    CasesPhotoContent:
      type: object
      required:
        - media_type
        - content_base64
      properties:
        media_type:
          type: string
          description: >-
            The media type storage's probe assigned to the bytes at complete
            time (the sanitized form), e.g. image/jpeg.
        content_base64:
          type: string
          description: >-
            The photo's stored bytes, base64-encoded -- the value a case view
            decodes into a blob URL it renders.
      description: The photo-content route's 200 answer.
    CasesError:
      type: object
      description: >-
        The structured {code, params} error envelope every speed API returns
        instead of localized text: code is a stable machine-readable identifier
        naming exactly one failure, and params carries the structured per-code
        details when the failure has any. The server never renders a message --
        a client maps each code to its own locale's text.
      properties:
        code:
          type: string
          example: cases.patient_name_required
        params:
          type: object
          additionalProperties: true
...